S_scan_const() Properly test if need to grow
As we parse the input, creating a string constant, we may have to grow
the destination if it fills up as we go along.  It allocates space in an
SV and populates the string, but it doesn' update the SvCUR until the
end, so in single stepping the debugger through the code, the SV looks
empty until the end.  It turns out that as a result SvEND also doesn't
get updated and still points to the beginning of the string until SvCUR
is finally set.  That means that the test changed by this commit was
always succeeding, because it was using SvEND that didn't get updated,
so it would attempt to grow each time through the loop.  By moving a
couple of statements earlier, and using SvLEN instead, which does always
have the correct value, those extra growth attempts are avoided.