S_scan_const: Make sure room for NUL in dest
At the end of  constant, we add a trailing NUL.  This makes sure there's
room for it.  But the code earlier was supposed to have already mad
enough space, so its a bug if there isn't enough space.  So on DEBUGGING
builds, we panic, as we've done before.  But otherwise we can continue
on with no harm having been done.